Changed int to size_t for calls to fread. This should never be a problem, but it is more correct.

+++ src/tttr.c
@@ -1,7 +1,7 @@
 #include "tttr.h"
 #include "error.h"
 
-void tttr_marker_print(FILE *stream_out, int marker) {
-	warn("External marker: %d\n", marker);
+void tttr_marker_print(FILE *stream_out, int64_t marker) {
+	warn("External marker: %"PRId64"\n", marker);
 }
